This is the first time when merge was done based on v2020.01 rebase tree
which contains patches which are out of tree. This branch shows how
transition happened.
Keep in your mind that final xlnx_rebase_v2020.01 branch can look
differently.

+++ cmd/bmp.c
@@ -9,16 +9,16 @@
  */
 
 #include <common.h>
-#include <dm.h>
-#include <lcd.h>
-#include <mapmem.h>
 #include <bmp_layout.h>
 #include <command.h>
-#include <asm/byteorder.h>
+#include <dm.h>
+#include <gzip.h>
+#include <lcd.h>
 #include <malloc.h>
 #include <mapmem.h>
 #include <splash.h>
 #include <video.h>
+#include <asm/byteorder.h>
 
 static int bmp_info (ulong addr);
 
@@ -57,7 +57,7 @@ struct bmp_image *gunzip_bmp(unsigned long addr, unsigned long *lenp,
 	bmp = dst;
 
 	/* align to 32-bit-aligned-address + 2 */
-	bmp = (struct bmp_image *)((((unsigned int)dst + 1) & ~3) + 2);
+	bmp = (struct bmp_image *)((((uintptr_t)dst + 1) & ~3) + 2);
 
 	if (gunzip(bmp, CONFIG_SYS_VIDEO_LOGO_MAX_SIZE, map_sysmem(addr, 0),
 		   &len) != 0) {
